refactoring and moving types
