+++ src/commands/_types.ts
@@ -1,16 +1,1 @@
-import type {
-  RESTPostAPIChatInputApplicationCommandsJSONBody,
-  SlashCommandBuilder,
-  SlashCommandSubcommandsOnlyBuilder,
-  SlashCommandOptionsOnlyBuilder,
-  ChatInputCommandInteraction,
-} from 'discord.js';
-
-export interface SlashCommand {
-  data:
-    | RESTPostAPIChatInputApplicationCommandsJSONBody
-    | SlashCommandSubcommandsOnlyBuilder
-    | Omit<SlashCommandBuilder, 'addSubcommandGroup' | 'addSubcommand'>
-    | SlashCommandOptionsOnlyBuilder;
-  execute: (interaction: ChatInputCommandInteraction) => Promise<unknown>;
-}
+export type { SlashCommand } from '../types/commands.js';
