+++ src/config.ts
@@ -1,32 +1,7 @@
 import 'dotenv/config';
 import { z } from 'zod';
 
-/**
- * Shape of the expected environment variables. This is used for type
- * inference only; the actual validation is done by zod below.
- */
-export interface ENV_SCHEMA {
-  DISCORD_TOKEN: string;
-  DISCORD_CLIENT_ID: string;
-  DISCORD_GUILD_ID?: string;
-  LOG_LEVEL: 'fatal' | 'error' | 'warn' | 'info' | 'debug' | 'trace' | 'silent';
-
-  N8N_SEARCH_URL?: string;
-  N8N_PLANT_URL?: string;
-  N8N_API_KEY?: string;
-
-  JELLYSEERR_URL?: string;
-  JELLYSEERR_API_KEY?: string;
-  JELLYSEERR_SERIES_DEFAULT: 'all' | 'first' | 'latest';
-  JELLYSEERR_4K: 'true' | 'false';
-  JELLYSEERR_AUTO_APPROVE: boolean;
-  JELLYSEERR_AUTO_DOWNLOAD: boolean;
-  JELLYSEERR_SEARCH_NOW: boolean;
-
-  QBIT_URL?: string;
-  QBIT_USERNAME?: string;
-  QBIT_PASSWORD?: string;
-}
+import type { EnvSchema } from './types/env.js';
 
 /**
  * Validate and parse environment variables using zod. Any required variables
@@ -65,4 +40,4 @@ if (!parsed.success) {
   process.exit(1);
 }
 
-export const env: ENV_SCHEMA = parsed.data as ENV_SCHEMA;
+export const env: EnvSchema = parsed.data as EnvSchema;
