+++ src/types/loggedFetch.ts
@@ -1,0 +1,7 @@
+export interface FetchLogMeta {
+  method: string;
+  url: string;
+  body?: string;
+  durationMs?: number;
+  status?: number;
+}
