+++ src/types/qbittorrent.ts
@@ -1,0 +1,7 @@
+export interface QbittorrentTorrent {
+  name: string;
+  progress: number;
+  dlspeed: number;
+  eta: number;
+  state: string;
+}
